+++ powerline_shell/__init__.py
@@ -6,22 +6,36 @@ import os
 import sys
 import importlib
 import json
-from .utils import warn, py3
+from .utils import warn, py3, import_file
 import re
 
 
+def _current_dir():
+    """Returns the full current working directory as the user would have used
+    in their shell (ie. without following symbolic links).
+
+    With the introduction of Bash for Windows, we can't use the PWD environment
+    variable very easily. `os.sep` for windows is `\` but the PWD variable will
+    use `/`. So just always use the `os` functions for dealing with paths. This
+    also is fine because the use of PWD below is done to avoid following
+    symlinks, which Windows doesn't have.
+
+    For non-Windows systems, prefer the PWD environment variable. Python's
+    `os.getcwd` function follows symbolic links, which is undesirable."""
+    if os.name == "nt":
+        return os.getcwd()
+    return os.getenv("PWD") or os.getcwd()
+
+
 def get_valid_cwd():
-    """ We check if the current working directory is valid or not. Typically
-        happens when you checkout a different branch on git that doesn't have
-        this directory.
-        We return the original cwd because the shell still considers that to be
-        the working directory, so returning our guess will confuse people
-    """
-    # Prefer the PWD environment variable. Python's os.getcwd function follows
-    # symbolic links, which is undesirable. But if PWD is not set then fall
-    # back to this func
+    """Determine and check the current working directory for validity.
+
+    Typically, an directory arises when you checkout a different branch on git
+    that doesn't have this directory. When an invalid directory is found, a
+    warning is printed to the screen, but the directory is still returned
+    as-is, since this is what the shell considers to be the cwd."""
     try:
-        cwd = os.getenv('PWD') or os.getcwd()
+        cwd = _current_dir()
     except:
         warn("Your current directory is invalid. If you open a ticket at " +
             "https://github.com/milkbikis/powerline-shell/issues/new " +
@@ -36,7 +50,7 @@ def get_valid_cwd():
         up = os.sep.join(parts)
     if cwd != up:
         warn("Your current directory is invalid. Lowest valid directory: "
-            + up)
+             + up)
     return cwd
 
 
@@ -100,7 +114,9 @@ class Powerline(object):
     def bgcolor(self, code):
         return self.color('48', code)
 
-    def append(self, content, fg, bg, separator=None, separator_fg=None):
+    def append(self, content, fg, bg, separator=None, separator_fg=None, sanitize=True):
+        if self.args.shell == "bash" and sanitize:
+            content = re.sub(r"([`$])", r"\\\1", content)
         self.segments.append((content, fg, bg,
             separator if separator is not None else self.separator,
             separator_fg if separator_fg is not None else bg))
@@ -115,16 +131,12 @@ class Powerline(object):
 
     def draw_segment(self, idx):
         segment = self.segments[idx]
-        if self.args.shell == "bash":
-            sanitized = re.sub(r"([`$])", r"\\\1", segment[0])
-        else:
-            sanitized = segment[0]
         next_segment = self.segments[idx + 1] if idx < len(self.segments)-1 else None
 
         return ''.join((
             self.fgcolor(segment[1]),
             self.bgcolor(segment[2]),
-            sanitized,
+            segment[0],
             self.bgcolor(next_segment[2]) if next_segment else self.reset,
             self.fgcolor(segment[4]),
             segment[3]))
@@ -134,6 +146,7 @@ def find_config():
     for location in [
         "powerline-shell.json",
         "~/.powerline-shell.json",
+        os.path.join(os.environ.get("XDG_CONFIG_HOME", "~/.config"), "powerline-shell", "config.json"),
     ]:
         full = os.path.expanduser(location)
         if os.path.exists(full):
@@ -154,6 +167,28 @@ DEFAULT_CONFIG = {
 }
 
 
+class ModuleNotFoundException(Exception):
+    pass
+
+
+class CustomImporter(object):
+    def __init__(self):
+        self.file_import_count = 0
+
+    def import_(self, module_prefix, module_or_file, description):
+        try:
+            mod = importlib.import_module(module_prefix + module_or_file)
+        except ImportError:
+            try:
+                module_name = "_custom_mod_{0}".format(self.file_import_count)
+                mod = import_file(module_name, os.path.expanduser(module_or_file))
+                self.file_import_count += 1
+            except (ImportError, IOError):
+                msg = "{0} {1} cannot be found".format(description, module_or_file)
+                raise ModuleNotFoundException( msg)
+        return mod
+
+
 def main():
     arg_parser = argparse.ArgumentParser()
     arg_parser.add_argument('--generate-config', action='store_true',
@@ -172,19 +207,33 @@ def main():
     config_path = find_config()
     if config_path:
         with open(config_path) as f:
-            config = json.loads(f.read())
+            try:
+                config = json.loads(f.read())
+            except Exception as e:
+                warn("Config file ({0}) could not be decoded! Error: {1}"
+                     .format(config_path, e))
+                config = DEFAULT_CONFIG
     else:
         config = DEFAULT_CONFIG
 
-    theme_name = config.get("theme", "default")
-    mod = importlib.import_module("powerline_shell.themes." + theme_name)
-    theme = getattr(mod, "Color")
+    custom_importer = CustomImporter()
+    theme_mod = custom_importer.import_(
+        "powerline_shell.themes.",
+        config.get("theme", "default"),
+        "Theme")
+    theme = getattr(theme_mod, "Color")
 
     powerline = Powerline(args, config, theme)
     segments = []
-    for seg_name in config["segments"]:
-        mod = importlib.import_module("powerline_shell.segments." + seg_name)
-        segment = getattr(mod, "Segment")(powerline)
+    for seg_conf in config["segments"]:
+        if not isinstance(seg_conf, dict):
+            seg_conf = {"type": seg_conf}
+        seg_name = seg_conf["type"]
+        seg_mod = custom_importer.import_(
+            "powerline_shell.segments.",
+            seg_name,
+            "Segment")
+        segment = getattr(seg_mod, "Segment")(powerline, seg_conf)
         segment.start()
         segments.append(segment)
     for segment in segments: