+++ powerline_shell/segments/time.py
@@ -6,12 +6,15 @@ import time
 class Segment(BasicSegment):
     def add_to_powerline(self):
         powerline = self.powerline
-        if powerline.args.shell == 'bash':
+        format = powerline.segment_conf('time', 'format')
+        if format:
+            time_ = ' %s ' % time.strftime(format)
+        elif powerline.args.shell == 'bash':
             time_ = ' \\t '
         elif powerline.args.shell == 'zsh':
             time_ = ' %* '
         else:
             time_ = ' %s ' % time.strftime('%H:%M:%S')
         powerline.append(time_,
-                         powerline.theme.HOSTNAME_FG,
-                         powerline.theme.HOSTNAME_BG)
+                         powerline.theme.TIME_FG,
+                         powerline.theme.TIME_BG)