+++ powerline_shell/themes/basic.py
@@ -1,4 +1,4 @@
-from .default import DefaultColor
+from powerline_shell.themes.default import DefaultColor
 
 
 class Color(DefaultColor):
@@ -40,3 +40,6 @@ class Color(DefaultColor):
 
     AWS_PROFILE_FG = 14
     AWS_PROFILE_BG = 8
+
+    TIME_FG = 8
+    TIME_BG = 7