+++ powerline_shell/utils.py
@@ -1,11 +1,17 @@
 import sys
+import os
 import threading
 
 py3 = sys.version_info[0] == 3
 
 if py3:
-    def unicode(x):
+    def unicode_(x):
         return str(x)
+    def decode(x):
+        return x.decode("utf-8")
+else:
+    unicode_ = unicode
+    decode = unicode
 
 
 class RepoStats(object):
@@ -16,7 +22,13 @@ class RepoStats(object):
         'staged': u'\u2714',
         'changed': u'\u270E',
         'new': u'?',
-        'conflicted': u'\u273C'
+        'conflicted': u'\u273C',
+        'stash': u'\u2398',
+        'git': u'\uE0A0',
+        'hg': u'\u263F',
+        'bzr': u'\u2B61\u20DF',
+        'fossil': u'\u2332',
+        'svn': u'\u2446'
     }
 
     def __init__(self, ahead=0, behind=0, new=0, changed=0, staged=0, conflicted=0):
@@ -61,7 +73,7 @@ class RepoStats(object):
 
             segment = repo_stats.n_or_empty("new") + icon_string
         """
-        return unicode(self[_key]) if int(self[_key]) > 1 else u''
+        return unicode_(self[_key]) if int(self[_key]) > 1 else u''
 
     def add_to_powerline(self, powerline):
         def add(_key, fg, bg):
@@ -82,14 +94,59 @@ def warn(msg):
 
 
 class BasicSegment(object):
-    def __init__(self, powerline):
+    def __init__(self, powerline, segment_def):
         self.powerline = powerline
+        self.segment_def = segment_def  # type: dict
 
     def start(self):
         pass
 
 
 class ThreadedSegment(threading.Thread):
-    def __init__(self, powerline):
+    def __init__(self, powerline, segment_def):
         super(ThreadedSegment, self).__init__()
         self.powerline = powerline
+        self.segment_def = segment_def  # type: dict
+
+
+def import_file(module_name, path):
+    # An implementation of https://stackoverflow.com/a/67692/683436
+    if py3 and sys.version_info[1] >= 5:
+        import importlib.util
+        spec = importlib.util.spec_from_file_location(module_name, path)
+        if not spec:
+            raise ImportError()
+        mod = importlib.util.module_from_spec(spec)
+        spec.loader.exec_module(mod)
+        return mod
+    elif py3:
+        from importlib.machinery import SourceFileLoader
+        return SourceFileLoader(module_name, path).load_module()
+    else:
+        import imp
+        return imp.load_source(module_name, path)
+
+
+def get_PATH():
+    """Normally gets the PATH from the OS. This function exists to enable
+    easily mocking the PATH in tests.
+    """
+    return os.getenv("PATH")
+
+
+def get_subprocess_env(**envs):
+    defaults = {
+        # https://github.com/milkbikis/powerline-shell/pull/153
+        "PATH": get_PATH(),
+    }
+    defaults.update(envs)
+    env = dict(os.environ)
+    env.update(defaults)
+    return env
+
+
+def get_git_subprocess_env():
+    # LANG is specified to ensure git always uses a language we are expecting.
+    # Otherwise we may be unable to parse the output.
+    return get_subprocess_env(LANG="C")
+